+++ vscode-ext/glyph-middleware.js
@@ -1509,7 +1509,7 @@ class GlyphCompressor {
 
   _createSourceMap() {
     return {
-      version: '1.33.1',
+      version: '1.33.2',
       level: this.level,
       provider: this.provider,
       profile: this.providerProfile,